Write to S3 using directory-like structure
This file uploads from a local ~/data/data folder on the user's system and iteratively writes every file in that directory into the specified S3
bucket. The structure in the bucket is presented (through the console) similarly to the file structure seen on the originating system. This takes
quite a while to iterate through large directories, so it is recommended to test by using a small sample set.